refactor: separate user schema to avoid model re-definition
