+++ users/db.js
@@ -1,5 +1,5 @@
 const mongoose = require("mongoose");
-
+const User = require("./schema");
 const mongoUri = process.env.MONGO_URI || "mongodb://mongo:27017/yovi";
 
 const seedUsers = [
@@ -13,7 +13,6 @@ async function connectDB() {
     console.log("MongoDB connected:", mongoUri);
 
     if (process.env.NODE_ENV !== "production") {
-      const User = mongoose.model("User");
       await User.deleteMany();
       await User.insertMany(seedUsers);
       console.log("Test data inserted");
